Allow user to specify /etc/resolv.conf target
Also add a Molecule scenario that tests this functionality.
