+++ molecule/specify_resolv_conf_target/INSTALL.rst
@@ -1,0 +1,1 @@
+../default/INSTALL.rst
