+++ molecule/specify_resolv_conf_target/requirements.yml
@@ -1,0 +1,1 @@
+../default/requirements.yml
