+++ molecule/specify_resolv_conf_target/tests/test_default_basic.py
@@ -1,0 +1,1 @@
+../../default/tests/test_default_basic.py
