+++ vars/default.yml
@@ -1,0 +1,7 @@
+---
+# The directory where systemd-resolved stores the _dynamic_
+# resolv.conf symlink targets it provides.
+dynamic_resolv_conf_target_dir: /run/systemd/resolve
+# The directory where systemd-resolved stores the _static_ resolv.conf
+# symlink targets it provides.
+static_resolv_conf_target_dir: /usr/lib/systemd
